Add ability to autofocus the window
Got tired of having to focus the window myself every time. On Linux you
can use the scroll wheel without the window having the focus, but I
wasn't able to find a way to send the keys to an unfocused Chrome window
on Windows, so autofocus is the second best option.