Fix MouseOnEvent UDF not firing mouse wheel events
In the Raw input handler, the mouse move event was overwriting the mouse
wheel event, making it so that if a user moves a mouse when turning the
mouse wheel, the mouse wheel event wouldn't fire. Removing the mouse
move event entirely fixes that, which is fine as we don't care about it
anyway.