Add an option to undo the autofocus
This is an attempt to imitate the behavior of being able to scroll tabs
in an inactive Chrome window, similar to how it is possible in Linux's
Chromium.

It appears as Chrome on Windows blocks all key input while the window is
inactive, so we are unable to send the next/previous tab shortcuts to it
unless we activate it. Activating a window brings it forward, which is
undesirable if we want to imitate the Linux behavior, so what we do is
activate it without bringing it forward. It is impossible to activate a
window without bringing it forward, but there is a trick we can use to
have it brought forward without it actually changing its Z-height
position: using the always-on-top window flag.

It's a bit hacky, but what happens is:

1. We collect information on all of the windows

2. All windows drawn above Chrome temporarily become always-on-top
   windows

3. All always-on-top windows are then made always-on-top again, forcing
   them to be above the windows in (2)

(Steps (2) and (3) are all done simultaneously, within a single screen
refresh cycle, but in that particular order, so the user shouldn't see
any visual changes)

4. Chrome window is activated. It is brought forward, but it can't be
   brought above the always-on-top windows, so visually it remains at
   the same Z-height, the only visual difference is that Chrome now has
   focus, instead of whatever window had it before.

5. Keys are sent to Chrome

6. Temporarily made always-on-top windows have their always-on-top flag
   removed

7. The previously active window gets focus back

While this works great, occasionally a window in-between the topmost
active window and Chrome gets drawn on top of the topmost active window
for a split screen refresh cycle, e.g. I have tested this with a Chrome
Incognito window in-between. Not sure what is causing this, but since it
happens rather rarely and gets fixed after that split refresh cycle,
that's a minor annoyance at the most.

Working on this feature has uncovered the fact that onMouseWheel() can
indirectly recurse into itself from within AutoIt function calls. This
has been a source of some head scratching, as the feature didn't behave
the way it should and I couldn't understand why. A queue has been added,
such that all recursion levels add events to the queue but only the
first recursion level processes the event queue.