Remove Hook mouse capture method and MouseOnEvent UDF
It looks like the Raw Input method of capturing the mouse input is
superior to hooking in every way, so the Hook method was removed. Also,
with how easy it is to setup Raw Input capture (following AutoIt
examples), it makes no sense to use MouseOnEvent UDF anymore.

Because Raw Input doesn't seem to affect the mouse performance, the
README was edited to reflect this and the "Gaming Mode" part was dropped
from the Disable tray menu button text.