Prevent page contents from scrolling
Occasionally a page would scroll, which should only happen if the mouse
is over the page contents so that it would receive the scroll event.
Perhaps it has something to do with a tab preview pop-up, or perhaps us
sending key presses to Chrome_RenderWidgetHostHWND somehow allows scroll
events to be received by the page. Whatever is the case, it looks like
moving the mouse in place, reminding Chrome that it's not over the page
contents but over the tabs area, seems to prevent the page from
scrolling. Or at least mostly prevent, as in some testing it would still
scroll, but very rarely, which is a big improvement.