Get the actual tabs area bounding box
Fallback to the hardcoded values if we fail to get it. It's unlikely to
fail though, in the worst case we would get a bounding box of a
different element, e.g. the entire window, which would be bad in its own
way and which we don't check for (how can we?).

Adjust some of the hardcoded tabs area offsets.

Simplify the logic for checking if a mouse cursor is in the tabs area,
as we don't need to check for left and top bounds with the mouse already
guaranteed to be within the window (the tabs area's left and top bounds
match those of the window).