Use Microsoft's UI Automation API to switch tabs
It's doesn't have all the issues we are having when sending keyboard
shortcuts: the shortcuts being interrupted by the tab preview popup
windows, the web page suddenly scrolling, etc. So it's a lot more
reliable in that sense. However, it makes the program more likely to
break due to a change in Chrome, as it depends on the UIA layout Chrome
has. I don't expect Chrome to change it's UIA layout that often for it
to be an issue, and even if it does change it -- it should be easy to
fix as I don't expect Chrome to remove all the information we need, it
would likely be just a tree structure change of sorts. The current code
is mostly resistant against such a change, with the tabs area bounding
box element selection being a hardcoded outlier.

It's very expensive to find which tab is currently selected (20-40ms),
however it's a lot faster to check if a specific tab is selected (<5ms).
To take advantage of this, 1) we note of the tab we select when
scrolling, as well as 2) subscribe to the tab change events from Chrome,
in case the user manually switches tabs without scrolling, caching them
((1) and (2)), so that when we need to scroll tabs we could check if any
of these two cached tabs are currently selected and go from there,
querying UIA about which tab is currently selected only as the last
resort. This caching makes the user experience very snappy. The last
resort fallback is needed because for some reason AutomationEventHandler
events tend to lag behind A LOT, especially if the Chrome window is
getting activated/deactivated (the whole UIA is slow on non-focused
windows), so it might not catch up in time if the user switches tabs too
fast. I wonder if this is an AutoIt issue and if we used C++ then events
wouldn't be so delayed?