Add a delay after focusing a window
Turns out that what was causing the scrolling of web pages wasn't the
mouse wheel but the PageUp and PageDown keys, i.e. the Ctrl key press
would somehow be lost and the browser would receive just the PgUp/PgDn.
It's obvious in retrospect, as a page would move just once and by more
than is possible in a single tick of a mouse wheel. Adding a delay after
focusing a Chrome window prevents the Ctrl key press from not being
delivered to it. The choice of the delay is not an obvious one, as I
would imagine that on slower computers it might require a longer delay
before a browser window becomes fully activated as for the browser to
accept key presses, so it was made into a user option.