Fix browser contents from occasionally scrolling
This finally fixes the long lasting issue for good.

We used to use Send() before, but after some changes and AutoIt update,
Send() was resulting in the Ctrl key being stuck in the pressed down
state. To work around this, we switched to using ControlSend(). However,
it turned out to have its own can of worms -- tab preview pop-ups were
eating up the keys we were trying to send, and the Ctrl key sometimes
was not being sent, with just PgUp/PgDn being sent, resulting in the web
page being scrolled instead of the tab being switched, which is
extremely annoying. While we worked around the tab preview pop-ups
eating keys, the page scrolling issue persisted. What made it especially
hard to troubleshoot is the inability to reproduce it -- it wasn't
obvious what exactly was causing it, it just happened occasionally. Only
recently I have managed to more or less reliably reproduce the page
scrolling issue: turns out it happens when you scroll just before or at
the time the tab preview pop-up gets created. The timing is very tight,
which is why when I have previously tested if the scrolling was caused
by the preview pop-ups I wasn't able to reproduce it. ControlSend()
attaches to the input thread of the application, so my guess is that
when a new control/window gets created while ControlSend() doing the
attachment, it somehow throws ControlSend() off, resulting in only
PgUp/PgDn being sent to the target control, so it's not something I can
fix without modifying AutoIt, which I can't do with it not being open
source. Recalling that none of this was the issue with the original
Send() code we had, I have switched to using that, and turns out you can
easily fix the issue of Ctrl being stuck by just pressing Ctrl once with
ControlSend() to unstuck it. So much time spent working around and
debugging ControlSend() issues when it was a lot easier to workaround
the Send() issue.