Add a more comprehensive autofocus undo option
The existing autofocus undo option is very fast, but it doesn't restore
window focus order. For example, you might have window1 and window2
above Chrome, so when you scroll, the focus changes from the currently
focused window2 to Chrome and then back to window2. While both window1
and window2 are still displayed on top of Chrome, this interaction
changes the window focus order. The focus order before scrolling was:
Chrome, window1 and then window2, but the interaction changed it to:
window1, Chrome, window2. This means that if you close window2, it's
Chrome that would become focused, and window1 would suddenly disappear
under Chrome even though it was above it just now. The expected behavior
would be for window1 to become focused instead. So the existing
autofocus undo option is kind of broken in a way, but I decided to keep
it because the proper alternative can be a lot slower, so there might be
cases when this broken version is actually desirable. It's also worth
noting that this seems to affect only (?) closing of the window2 -- if
you minimize it instead, then the window1 would become focused as
expected. It could be affecting more than just closing, but if so, I
have yet to discover that.

The proper comprehensive option is a lot slower, as after focusing
Chrome and sending the keys, it focuses the windows above Chrome one by
one, respecting the focus delay. That way the focus order is
restored/preserved. It's not particularly slow if there are only one or
two windows above Chrome, but if there are more of them, the focus delay
adds up and it becomes noticeably slower than the simple autofocus undo
option.