Add an option to restart as admin
Add a convenient option to restart the program as the Administrator,
instead of having to fish out the executable to do so. Also nudge the
user to run it as Administrator by making it the first tray menu item.
We want the user to run the program as an Administrator because it
improves the UX. If the currently focused window is a privileged window,
e.g. Task Manager, the user won't be able to scroll Chrome tabs behind
that window as we won't be receiving any mouse events due to the
privilege isolation. Still, not everyone might be able to run the
program as an Administrator, e.g. a student running it on a university
computer, and the program is plenty useful when running as a regular
user, which is why it's an option and not the executable manifest
requirement.