Add an option to run at startup
While it's easy enough for a user to setup autostart by adding a
shortcut to shell:startup, this will start the program only as the
regular user, it will not start the program elevated, i.e. as the
Administrator. It appears that on Windows one must use the Task
Scheduler to autostart a program with elevated rights, which is a bit
harder to setup than dropping a shortcut into shell:startup, so the
option automating this process was added. A user can also setup an
non-elevated startup, which the program does via the registry. Using the
registry seems to be the preferred method for when a software sets up
the autostart on its own, leaving the shell:execute method mostly for
user manually adding/removing programs to/from autostart.

An issue arises when a user wants to autostart the program with elevated
rights: if the program is stored in an insecure location, then
unprivileged applications could overwrite it, which will result in those
unprivileged applications being autorun with elevated rights on the next
startup -- a privilege escalation vulnerability. To workaround this
issue, the program prompts the user that it will move files to a known
secure location, the Program Files directory, so that non-elevated
programs couldn't overwrite it.

Setting up a task that would run as Administrator with Task Scheduler
and moving files to Program Files requires elevated rights, so the
script restarts with elevated rights when trying to setup the elevated
autostart. This might result in a somewhat amusing series of restarts:
Unprivileged process with the .exe in Desktop -> Privileged process with
the .exe in Desktop -> Privileged process with the .exe in Program
Files.

Setting up an autostart as a regular user doesn't require any elevated
rights. This option also supports roaming between machines, as user's
registry entries roam around, when Task Manager is machine-specific.
This is a useful option for when the user doesn't have administrative
rights on a computer, e.g. when using a program on a college computer.

The startup option tray items that require elevation lack the
corresponding shield icon next to them due to the tray items being radio
items -- the space that otherwise could be used for the shield icon is
taken up by the dot indicating which of the options is selected. This is
a bad UX, but not much we can do about it unless we move away from
having all the options in the tray menu and having a proper settings
GUI, which I'd like to avoid for now since there are only a handful of
options.